requirements.txt
torch
torchvision
numpy
matplotlib
pandas
Pillow
scikit-learn
